handle retarget better
git-svn-id: http://svn.vuze.com/public/client@40437 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/org/gudy/azureus2/core3/disk/impl/DiskManagerUtil.java
@@ -257,21 +257,33 @@ DiskManagerUtil
 	
 	        }else{
 	
-                if ( FileUtil.deleteWithRecycle( 
-            		existing_file,
-            		download_manager.getDownloadState().getFlag( DownloadManagerState.FLAG_LOW_NOISE ))){
-	
-	                    // new file, recheck
-	
-	                download_manager.recheckFile( file_info );
-	
-	            }else{
-	
-	                Logger.log(new LogAlert(download_manager, LogAlert.REPEATABLE, LogAlert.AT_ERROR,
-	                        "Failed to delete '" + existing_file.toString() + "'"));
-	
-	                return( false );
-	            }
+	        	Object	skip_delete = download_manager.getUserData( "set_link_dont_delete_existing" );
+	        	
+	        		// user is manually re-targetting a file - this hack is to stop us from deleting the old file which 
+	        		// we really have no right to go and delete
+	        	
+	        	if ( ( skip_delete instanceof Boolean ) && (Boolean)skip_delete ){
+	        		
+	        		 download_manager.recheckFile( file_info );
+	        		 
+	        	}else{
+	        		
+	                if ( FileUtil.deleteWithRecycle( 
+	            		existing_file,
+	            		download_manager.getDownloadState().getFlag( DownloadManagerState.FLAG_LOW_NOISE ))){
+		
+		                    // new file, recheck
+		
+		                download_manager.recheckFile( file_info );
+		
+		            }else{
+		
+		                Logger.log(new LogAlert(download_manager, LogAlert.REPEATABLE, LogAlert.AT_ERROR,
+		                        "Failed to delete '" + existing_file.toString() + "'"));
+		
+		                return( false );
+		            }
+	        	}
 	        }
 	    }else{
